https://github.com/pmocal/imaginaryinventory/issues/11

+++ controllers/categoryController.js
@@ -50,19 +50,21 @@ exports.category_create_post = [
 	validator.sanitizeBody('name').escape(),
 	//process request after validation & sanitization
 	(req, res, next) => {
-		const tempPath = req.file.path;
-		const targetPath = path.join(__dirname, "../public/uploads/" + req.body.name + ".png");
-		if (path.extname(req.file.originalname).toLowerCase() === ".png") {
-			fs.rename(tempPath, targetPath, err => {
-				if (err) return next(err);
-		  	});
-		} else {
-			fs.unlink(tempPath, err => {
-				if (err) return next(err);
-				res.status(403)
-					.contentType("text/plain")
-					.end("Only .png files are allowed!");
-			});
+		if (req.file != null) {
+			const tempPath = req.file.path;
+			const targetPath = path.join(__dirname, "../public/uploads/" + req.body.name + ".png");
+			if (path.extname(req.file.originalname).toLowerCase() === ".png") {
+				fs.rename(tempPath, targetPath, err => {
+					if (err) return next(err);
+			  	});
+			} else {
+				fs.unlink(tempPath, err => {
+					if (err) return next(err);
+					res.status(403)
+						.contentType("text/plain")
+						.end("Only .png files are allowed!");
+				});
+			}
 		}
 		//Extract the validation errors from a request
 		const errors = validator.validationResult(req);
@@ -160,12 +162,29 @@ exports.category_update_get = function(req, res, next) {
 }
 
 exports.category_update_post = [
+	upload.single('file'),
 	//validate that name field not empty
 	validator.body('name', 'Category name required').isLength({ min: 1 }).trim(),
 	//sanitize name field
 	validator.sanitizeBody('name').escape(),
 	//process request after validation & sanitization
 	(req, res, next) => {
+		if (req.file != null) {
+			const tempPath = req.file.path;
+			const targetPath = path.join(__dirname, "../public/uploads/" + req.body.name + ".png");
+			if (path.extname(req.file.originalname).toLowerCase() === ".png") {
+				fs.rename(tempPath, targetPath, err => {
+					if (err) return next(err);
+			  	});
+			} else {
+				fs.unlink(tempPath, err => {
+					if (err) return next(err);
+					res.status(403)
+						.contentType("text/plain")
+						.end("Only .png files are allowed!");
+				});
+			}
+		}
 		//Extract the validation errors from a request
 		const errors = validator.validationResult(req);
 		//create genre object with escaped & trimmed data
